r11679: opening/creating files in SMB2 now works. Lots of unknown parameters in the call tho.

+++ source/libcli/smb2/smb2.h
@@ -56,7 +56,7 @@ struct smb2_transport {
 */
 struct smb2_tree {
 	struct smb2_session *session;
-	uint64_t tid;
+	uint32_t tid;
 };
 
 /*
